When verifying soft links, do not attempt to delete the trailing slash from the current on-disk state if the expected link destination also has a trailing slash, proposed by Jonathan Chen

+++ src/files_links.c
@@ -114,7 +114,10 @@ if (readlink(destination,linkbuf,CF_BUFSIZE-1) == -1)
 else
    { int off1 = 0, off2 = 0;  /* Link exists */
    
-   DeleteSlash(linkbuf);
+   if (*to && !IsFileSep(to[strlen(to)-1]))
+      {
+      DeleteSlash(linkbuf);
+      }
    
    if (strncmp(linkbuf,"./",2) == 0)   /* Ignore ./ at beginning */
       {
